src/dashboard/requirements.txt
streamlit
plotly
kafka-python
pandas
hdfs
